Move some behavior from the views to the model, and all some docstrings.
